Unit test to demonstrate #15
In this unit test I create 2 concave polygons. These polygon data is
derrived from a path iterator used to create fonts. The 2 letters chosen
are a capital H and lower case e (just the outer shape, the inner hole
has been removed from this dataset to demonstrate a specific problem).
The 2 sets of points are extruded and the number of polygons displayed
of the 2. When a simple union of non intersecting CSG's is performed,
the total number of polygons should not change.