creative fix for stackoverflow issue
we're adjusting two things... still the root cause of the stackoverflow is still not clear.
we're using a random polygon to define the plane for the node. chances are bigger that find a a coplanar polygon.
then if we have only one polygon as an input, we're expecting it to be coplanar.

a better alternative  may be to say that as long as the polygon from which we defined the plane is not SPANNING, it must be COPLANAR. 

anyway, this fix made my situation work again. 

maybe the issues arise from bigger models somewhere in the 100 dimension